Moved bear routes into own file, set up bearsModel.js file. Fixed error when trying to update non-existant data.
