+++ bears/bearsRoutes.js
@@ -1,13 +1,13 @@
 const express = require('express');
 const router = express.Router();
 
-const zoos = require('./zoosModel.js');
+const bears = require('./bearsModel.js');
 
 router.get('/', (req, res) => {
-  zoos
+  bears
     .find()
-    .then((zoos) => {
-      res.status(200).json(zoos);
+    .then((bears) => {
+      res.status(200).json(bears);
     })
     .catch((err) => {
       res.status(500).json(err);
@@ -16,15 +16,15 @@ router.get('/', (req, res) => {
 
 router.get('/:id', (req, res) => {
   const { id } = req.params;
-  zoos
+  bears
     .findById(id)
-    .then((zoo) => {
-      if (!zoo) {
+    .then((bear) => {
+      if (!bear) {
         return res.status(404).send({
-          message: `The zoo with the specified ID ${id} does not exist.`,
+          message: `The bear with the specified ID ${id} does not exist.`,
         });
       }
-      res.status(200).json(zoo);
+      res.status(200).json(bear);
     })
     .catch((err) => {
       res.status(500).json(err);
@@ -33,14 +33,14 @@ router.get('/:id', (req, res) => {
 
 router.post('/', (req, res) => {
   const { name } = req.body;
-  const zoo = { name };
+  const bear = { name };
   if (!name) {
     return res.status(400).send({
-      errorMessage: 'Please provide a name for the zoo.',
+      errorMessage: 'Please provide a name for the bear.',
     });
   }
-  zoos
-    .add(zoo)
+  bears
+    .add(bear)
     .then((ids) => {
       res.status(201).json(ids);
     })
@@ -52,16 +52,16 @@ router.post('/', (req, res) => {
 router.put('/:id', (req, res) => {
   const { id } = req.params;
   const { name } = req.body;
-  const newZoo = { name };
+  const newbear = { name };
   if (!name) {
     return res.status(400).send({
-      errorMessage: 'Please provide a name for the zoo.',
+      errorMessage: 'Please provide a name for the bear.',
     });
   }
-  zoos.update(id, newZoo).then((count) => {
+  bears.update(id, newbear).then((count) => {
     if (!count) {
       return res.status(404).json({
-        message: 'No zoo found to update',
+        message: 'No bear found to update',
       });
     }
     res.status(200).json({ message: `${count} record updated` });
@@ -70,15 +70,15 @@ router.put('/:id', (req, res) => {
 
 router.delete('/:id', (req, res) => {
   const { id } = req.params;
-  zoos
+  bears
     .remove(id)
     .then((remove) => {
       if (!remove) {
         return res.status(404).send({
-          message: `The zoo with the specified ID ${id} does not exist.`,
+          message: `The bear with the specified ID ${id} does not exist.`,
         });
       }
-      res.status(200).send({ message: `zoo with ID ${id} was removed.` });
+      res.status(200).send({ message: `bear with ID ${id} was removed.` });
     })
     .catch((err) => {
       res.status(500).json(err);
